+++ akad/__init__.py
@@ -7,4 +7,4 @@ Public API::
 from akad.sdk import DataContractBreachError, DataContractValidator
 
 __all__ = ["DataContractValidator", "DataContractBreachError"]
-__version__ = "1.1.0"
+__version__ = "1.2.0"